+++ src/features/SideBar/index.tsx
@@ -1,6 +1,5 @@
 import { SideNav } from '@lobehub/ui';
 import { memo } from 'react';
-import { shallow } from 'zustand/shallow';
 
 import AvatarWithUpload from '@/features/AvatarWithUpload';
 import { useSettings } from '@/store/settings';
@@ -9,7 +8,7 @@ import BottomActions from './BottomActions';
 import TopActions from './TopActions';
 
 export default memo(() => {
-  const [tab, setTab] = useSettings((s) => [s.sidebarKey, s.switchSideBar], shallow);
+  const [tab, setTab] = useSettings((s) => [s.sidebarKey, s.switchSideBar]);
 
   return (
     <SideNav