+++ src/pages/chat/SessionList/List/index.tsx
@@ -1,20 +1,21 @@
 import isEqual from 'fast-deep-equal';
 import Link from 'next/link';
 import { memo } from 'react';
-import { shallow } from 'zustand/shallow';
 
-import { sessionSelectors, useSessionStore } from '@/store/session';
+import { sessionSelectors, useSessionHydrated, useSessionStore } from '@/store/session';
 
 import SessionItem from './SessionItem';
+import SkeletonItem from './SkeletonItem';
 
 const SessionList = memo(() => {
   const list = useSessionStore(sessionSelectors.sessionList, isEqual);
-  const [activeId, loading] = useSessionStore(
-    (s) => [s.activeId, s.autocompleteLoading.title],
-    shallow,
-  );
+  const [activeId, loading] = useSessionStore((s) => [s.activeId, s.autocompleteLoading.title]);
 
-  return (
+  const isInit = useSessionHydrated();
+
+  return !isInit ? (
+    <SkeletonItem />
+  ) : (
     <>
       {list.map(({ id }) => (
         <Link href={`/chat/${id}`} key={id}>